requirements.txt
Flask==0.10.1
#Flask-Sockets==0.1
git+https://github.com/jolleon/flask-sockets.git@master#egg=Flask-Sockets
Jinja2==2.7.1
MarkupSafe==0.18
Werkzeug==0.9.4
distribute==0.6.34
gevent==0.13.8
gevent-websocket==0.3.6
greenlet==0.4.1
gunicorn==18.0
itsdangerous==0.23
redis==2.8.0
wsgiref==0.1.2
docker-py==0.3.1
simplejson==3.4.1

